provide user a disconnect button after connecting
After a user has connected a myAlgo wallet the ui should remove the
original `connect` button and replace it with a `Disconnect` button.
This should help the user understand that the connect request was
successful and the application "knows" about the connected address.

The disconnect button allows the user to reset the application state by
setting myAddress back to an empty string.